gameProject refactor
