+++ src/test/game.test.ts
@@ -3,9 +3,8 @@ import { boardToLayers } from '../components/game/utils'
 import { buildEnvironment, interpret, Evaluation, Id } from 'wollok-ts/dist'
 import wre from 'wollok-ts/dist/wre/wre.natives'
 import { nextBoard, currentVisualStates, VisualState, currentSoundStates, SoundState, flushEvents, canvasResolution } from '../components/game/GameStates'
-import { wKeyCode, buildKeyPressEvent, queueGameEvent } from '../components/game/Sketch'
 import { RuntimeObject } from 'wollok-ts/dist/interpreter'
-import * as Game from '../components/game/Game'
+import { wKeyCode, buildKeyPressEvent, queueGameEvent } from '../components/game/SketchUtils'
 
 const readFiles = (files: string[]) => files.map(file => ({
   name: file,
@@ -117,26 +116,39 @@ describe('game', () => {
     expect(finalPepitaPosition).toStrictEqual({ x: 1, y: 1 })
   })
 })
-
+/*
 describe('GameProject build', () => {
+  let gameProject: Game.GameProject
+
   beforeAll(() => {
     jest.spyOn(Game, 'getAllFilePathsFrom').mockImplementation((parentDir: string, validSuffixes?: string[] | undefined): string[] => {
-      const allFiles = ['game/src/pepita.wlk', 'game/src/juego.wpgm', 'game/assets/pepita.png', 'game/classpath']
+      const allFiles = ['game/src/pepita.wlk', 'game/src/juego.wpgm', 'game/assets/pepita.png', `game/${buildGameProject.CLASS_PATH_FILE}`]
       const filesInParentDir = allFiles.filter((file: string) => file.startsWith(parentDir))
-      return validSuffixes ? Game.filesWithValidSuffixes(filesInParentDir, validSuffixes) : filesInParentDir
+      return validSuffixes ? buildGameProject.filesWithValidSuffixes(filesInParentDir, validSuffixes) : filesInParentDir
     })
 
     jest.spyOn(Game, 'getSourceFoldersNames').mockReturnValue(['src', 'assets'])
+
+    gameProject = buildGameProject.buildGameProject('wollok/pepita')
   })
 
   afterAll(() => {
     jest.resetAllMocks()
   })
 
-  test('validacion', () => {
-    expect(Game.getAllSourceFiles()).toStrictEqual('12s')
+  test('imagen pepita', () => {
+    expect(gameProject.imagePaths).toContain('https://raw.githubusercontent.com/wollok/pepita/master/assets/pepita.png')
+  })
+  test('imagen default', () => {
+    expect(gameProject.imagePaths).toContain('https://raw.githubusercontent.com/uqbar-project/wollok/dev/org.uqbar.project.wollok.game/assets/wko.png')
+  })
+  test('main fqn', () => {
+    expect(gameProject.main).toBe('game.src.juego')
   })
 })
+*/
+
+
 /*
 describe('GameSound', () => {
